Rewrite ModalManagerState into Typescript
- Fixes `attrs` parameter being marked as required
- Add `isModalOpen` method
